+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added: New validation `expressionDoesNotUseLabelsForMetricRegexp`
 
 ## [3.6.2] - 2025-02-07
 - Fixed: Invalid internal name of the `newExpressionDoesNotUseClassicHistogramBucketOperations` validator causing that it wasn't possible to disable the validator